Hoist queryClient and use shared current user hook
