+++ packages/user/FractalCore/ui/src/hooks/fractals/use-guild-membership.ts
@@ -3,7 +3,8 @@ import { useQuery } from "@tanstack/react-query";
 import { getGuildMembership } from "@/lib/graphql/fractals/getGuildMembership";
 import QueryKey, { OptionalAccount } from "@/lib/queryKeys";
 
-import { useCurrentUser } from "../use-current-user";
+import { useCurrentUser } from "@shared/hooks/use-current-user";
+
 import { useGuildAccount } from "../use-guild-account";
 
 export const useGuildMembership = (
