+++ packages/user/FractalCore/ui/src/hooks/fractals/use-set-schedule.ts
@@ -1,4 +1,3 @@
-import { queryClient } from "@/queryClient";
 import { useMutation } from "@tanstack/react-query";
 import { z } from "zod";
 
@@ -6,11 +5,12 @@ import QueryKey from "@/lib/queryKeys";
 import { zUnix } from "@/lib/zod/Unix";
 import { zGuildAccount } from "@/lib/zod/Wrappers";
 
+import { assertUser } from "@shared/hooks/use-current-user";
+import { queryClient } from "@shared/lib/queryClient";
 import { zAccount } from "@shared/lib/schemas/account";
 import { supervisor } from "@shared/lib/supervisor";
 import { toast } from "@shared/shadcn/ui/sonner";
 
-import { assertUser } from "../use-current-user";
 import { setDefaultMembership } from "./use-membership";
 
 const zParams = z.object({
