--- packages/user/FractalCore/ui/src/hooks/use-current-user.ts
@@ -1,42 +1,0 @@
-import { queryClient } from "@/queryClient";
-import { useQuery } from "@tanstack/react-query";
-import { z } from "zod";
-
-import QueryKey from "@/lib/queryKeys";
-
-import { Account, zAccount } from "@shared/lib/schemas/account";
-import { supervisor } from "@shared/lib/supervisor";
-
-export type GetCurrentUserRes = string | null;
-
-export const queryFn = async () => {
-    try {
-        const res = await supervisor.functionCall({
-            method: "getCurrentUser",
-            params: [],
-            service: "accounts",
-            intf: "api",
-        });
-        return res ? z.string().parse(res) : null;
-    } catch (error) {
-        const message = "Error getting current user";
-        console.error(message, error);
-        throw new Error(message);
-    }
-};
-
-export const useCurrentUser = (refetchInterval?: number) =>
-    useQuery({
-        queryKey: QueryKey.currentUser(),
-        queryFn,
-        refetchInterval,
-        staleTime: 60000,
-    });
-
-export const assertUser = (): Account =>
-    zAccount.parse(queryClient.getQueryData(QueryKey.currentUser()));
-
-export const getCurrentUser = (): string | null => {
-    const res = queryClient.getQueryData(QueryKey.currentUser());
-    return res ? z.string().parse(res) : null;
-};
