+++ packages/user/FractalCore/ui/src/main.tsx
@@ -9,9 +9,9 @@ import { createRoot } from "react-dom/client";
 import { RouterProvider } from "react-router-dom";
 
 import { ThemeProvider } from "@shared/components/theme-provider";
+import { queryClient } from "@shared/lib/queryClient";
 import { Toaster } from "@shared/shadcn/ui/sonner";
 
-import { queryClient } from "./queryClient";
 import { router } from "./router";
 
 dayjs.extend(advancedFormat);
