+++ packages/user/FractalCore/ui/src/pages/fractal/branches/judicial/index.tsx
@@ -5,10 +5,10 @@ import { GuildOverviewCard } from "@/components/guild-overview-card";
 import { ExileFractalMemberModal } from "@/components/modals/exile-fractal-member-modal";
 
 import { useFractal } from "@/hooks/fractals/use-fractal";
-import { useCurrentUser } from "@/hooks/use-current-user";
 import { useGuild } from "@/hooks/use-guild";
 
 import { ErrorCard } from "@shared/components/error-card";
+import { useCurrentUser } from "@shared/hooks/use-current-user";
 import { Button } from "@shared/shadcn/ui/button";
 import {
     Card,
